refactor(commands): split chat transport
